IGNITE-23830 Remove forceServerMode flag
