+++ modules/core/src/test/java/org/apache/ignite/internal/processors/cache/distributed/IgniteCacheNearRestartRollbackSelfTest.java
@@ -41,7 +41,6 @@ import org.apache.ignite.configuration.NearCacheConfiguration;
 import org.apache.ignite.internal.IgniteInternalFuture;
 import org.apache.ignite.lang.IgniteFuture;
 import org.apache.ignite.spi.communication.tcp.TcpCommunicationSpi;
-import org.apache.ignite.spi.discovery.tcp.TcpDiscoverySpi;
 import org.apache.ignite.testframework.GridTestUtils;
 import org.apache.ignite.testframework.junits.common.GridCommonAbstractTest;
 import org.apache.ignite.transactions.Transaction;
@@ -68,9 +67,6 @@ public class IgniteCacheNearRestartRollbackSelfTest extends GridCommonAbstractTe
 
         cfg.setClientFailureDetectionTimeout(50000);
         cfg.setCacheConfiguration(cacheConfiguration(igniteInstanceName));
-
-        ((TcpDiscoverySpi)cfg.getDiscoverySpi())
-            .setForceServerMode(getTestIgniteInstanceName(3).equals(igniteInstanceName));
 
         TcpCommunicationSpi commSpi = new TcpCommunicationSpi();
